+++ auth0.go
@@ -2,7 +2,6 @@ package auth0
 
 import (
 	"fmt"
-	"strconv"
 	"time"
 )
 
@@ -16,45 +15,6 @@ func BoolValue(b *bool) bool {
 		return *b
 	}
 	return false
-}
-
-// ConvertibleBoolean is a custom bool type that is able to unmarshal/marshal from both string and bool types
-type ConvertibleBoolean bool
-
-// ConvertibleBool returns a pointer to the ConvertibleBoolean value passed in.
-func ConvertibleBool(b bool) *ConvertibleBoolean {
-	cb := ConvertibleBoolean(b)
-	return &cb
-}
-
-// ConvertibleBoolValue returns the value of the ConvertibleBoolean pointer passed in or false if the
-// pointer is nil.
-func ConvertibleBoolValue(b *ConvertibleBoolean) ConvertibleBoolean {
-	if b != nil {
-		return *b
-	}
-	return false
-}
-
-// UnmarshalJSON handles unmarshalling of the ConvertibleBoolean value in the case of string and also bool values
-func (bit ConvertibleBoolean) UnmarshalJSON(data []byte) error {
-	asString := trimQuotes(string(data))
-	b, err := strconv.ParseBool(asString)
-	if err != nil {
-		return err
-	}
-	bit = ConvertibleBoolean(b)
-	return nil
-}
-
-func trimQuotes(s string) string {
-	if len(s) > 0 && s[0] == '"' {
-		s = s[1:]
-	}
-	if len(s) > 0 && s[len(s)-1] == '"' {
-		s = s[:len(s)-1]
-	}
-	return s
 }
 
 // Int returns a pointer to the int value passed in.
