+++ auth0_test.go
@@ -21,21 +21,6 @@ func TestBool(t *testing.T) {
 	}
 }
 
-func TestConvertibleBool(t *testing.T) {
-	for _, test := range []struct {
-		in       *ConvertibleBoolean
-		expected ConvertibleBoolean
-	}{
-		{nil, false},
-		{ConvertibleBool(false), false},
-		{ConvertibleBool(true), true},
-	} {
-		have := ConvertibleBoolValue(test.in)
-		if have != test.expected {
-			t.Errorf("unexpected output. have %v, expected %v", have, test.expected)
-		}
-	}
-}
 
 func TestInt(t *testing.T) {
 	for _, test := range []struct {
